väike loogika bugi parandatud ja broadcastMessage liigutatud ServerConnection'isse. serveri Main liigutatud otse ServerConnection klassi, mis on nüüd ServerMain klass
